+++ bolero/behavior_search/__init__.py
@@ -1,7 +1,8 @@
 from .behavior_search import BehaviorSearch, ContextualBehaviorSearch
 from .black_box_search import (BlackBoxSearch, ContextualBlackBoxSearch,
                                JustOptimizer, JustContextualOptimizer)
+from .monte_carlo_rl import MonteCarloRL
 
-__all__ = ["BehaviorSearch", "ContextualBehaviorSearch",
-           "BlackBoxSearch", "ContextualBlackBoxSearch",
-           "JustOptimizer", "JustContextualOptimizer"]
+__all__ = ["BehaviorSearch", "ContextualBehaviorSearch", "BlackBoxSearch",
+           "ContextualBlackBoxSearch", "JustOptimizer",
+           "JustContextualOptimizer", "MonteCarloRL"]