+++ bolero/environment/openaigym.py
@@ -129,7 +129,7 @@ class OpenAiGym(Environment):
     def step_action(self):
         inputs = self.input_handler(self.inputs)
         observations, reward, done, info = self.env.step(inputs)
-        self.outputs[:] = observations.ravel()
+        self.outputs[:] = np.atleast_1d(observations).ravel()
         self.rewards.append(reward)
         self.done = self.done or done
 
@@ -156,3 +156,18 @@ class OpenAiGym(Environment):
             return np.inf
         else:
             return self.env.spec.reward_threshold
+
+    def get_discrete_action_space(self):
+        """Get list of possible actions.
+
+        An error will be raised if the action space of the problem is not
+        discrete. The environment must be initialized before this method can
+        be called.
+
+        Returns
+        -------
+        """
+        if not hasattr(self.env.action_space, "n"):
+            raise TypeError("gym environment '%d' does not have a discrete "
+                            "action space")
+        return list(range(self.env.action_space.n))